Updated timezone to use an enumeration as opposed to a numeric input. This is translated but more user-friendly,
Fixed issue with days of week being incorrectly interpreted for sessions if there is a cross-over after server-time adjustment.
Optimized session time indicator to only add if being used otherwise it is ignored.
Input parameters updated to accomodate user-friendly input but by default it will be empty so must be reset in your input parameters.